refactor: address clippy lints and implement Default
- Implement `Default` for `PullCancelToken`.
- Simplify `candidates.iter().any(...)` to `candidates.contains(...)` in `whatsapp_pull.rs`.
- Remove unnecessary `as usize` cast in `db/mod.rs`.
- Simplify bitwise operation in `whatsapp/mod.rs` test helper.

+++ src-tauri/src/db/mod.rs
@@ -256,7 +256,7 @@ pub fn move_files_to_trash_with_connection(
         removed_from_index += connection.execute(
             "DELETE FROM files WHERE absolute_path = ?1",
             params![path],
-        )? as usize;
+        )?;
     }
 
     Ok(TrashResult {
